keras_tf_requirements.txt
numpy
opencv-contrib-python
scipy
matplotlib
pillow
imutils
h5py
requests
progressbar2
scikit-learn
scikit-image
tensorflow-gpu==1.14.0
keras